:art: fix: Member 도메인 model 패키지로 이동
