+++ src/main/java/com/notitime/noffice/api/member/business/MemberService.java
@@ -1,0 +1,4 @@
+package com.notitime.noffice.api.member.business;
+
+public class MemberService {
+}
