+++ src/main/java/com/notitime/noffice/api/member/presentation/MemberController.java
@@ -1,5 +1,6 @@
-package com.notitime.noffice.presentation;
+package com.notitime.noffice.api.member.presentation;
 
+import com.notitime.noffice.api.auth.business.AuthService;
 import com.notitime.noffice.global.response.BusinessSuccessCode;
 import com.notitime.noffice.global.response.NofficeResponse;
 import com.notitime.noffice.request.SocialAuthRequest;
@@ -24,17 +25,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class MemberController {
 
+	private final AuthService authService;
+
 	@Operation(summary = "회원 로그인", description = "인증 헤더에 토큰을, 본문에 소셜 로그인 정보를 넣어 노피스 서버 로그인을 시도합니다.")
 	@PostMapping("/login")
 	public NofficeResponse<SocialAuthResponse> login(@RequestHeader("Authorization") final String authorization,
 	                                                 @RequestBody final SocialAuthRequest socialLoginRequest) {
-		return NofficeResponse.success(BusinessSuccessCode.POST_LOGIN_SUCCESS);
+		return NofficeResponse.success(BusinessSuccessCode.POST_LOGIN_SUCCESS, authService.login(socialLoginRequest));
 	}
 
 	@Operation(summary = "토큰 재발급", description = "리프레시 토큰을 이용해 새로운 액세스 토큰을 발급합니다.")
 	@PostMapping("/reissue")
 	public NofficeResponse<TokenResponse> reissue(@RequestHeader("Authorization") final String refreshToken) {
-		return NofficeResponse.success(BusinessSuccessCode.POST_REISSUE_SUCCESS);
+		return NofficeResponse.success(BusinessSuccessCode.POST_REISSUE_SUCCESS, authService.reissue(refreshToken));
 	}
 
 	@Operation(summary = "단일 회원 정보 조회")
