--- src/main/java/com/notitime/noffice/domain/Member.java
@@ -1,31 +1,0 @@
-package com.notitime.noffice.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Member extends BaseTimeEntity {
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	private String name;
-
-	private String alias;
-
-	private String profileImage;
-
-	@OneToMany(mappedBy = "member")
-	private final List<OrganizationMember> organizations = new ArrayList<>();
-}
