+++ src/main/java/com/notitime/noffice/domain/OrganizationMember.java
@@ -1,6 +1,9 @@
 package com.notitime.noffice.domain;
 
+import com.notitime.noffice.domain.member.model.Member;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -19,6 +22,7 @@ public class OrganizationMember {
 	@GeneratedValue
 	private Long id;
 
+	@Enumerated(EnumType.STRING)
 	private OrganizationRole role;
 
 	@ManyToOne(fetch = FetchType.LAZY)
