+++ src/main/java/com/notitime/noffice/domain/TaskStatus.java
@@ -1,5 +1,6 @@
 package com.notitime.noffice.domain;
 
+import com.notitime.noffice.domain.member.model.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
