+++ src/main/java/com/notitime/noffice/domain/member/persistence/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.notitime.noffice.domain.member.persistence;
+
+import com.notitime.noffice.domain.member.model.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+	Optional<Member> findBySerialId(String serialId);
+}
